Avance del proyecto

+++ src/models/gormDB/plan.go
@@ -9,7 +9,7 @@ type Plan struct {
 	Aproximacion_baja     *uint32  `json:"aproximacion_baja,omitempty"`
 	Nombre                *string  `json:"nombre,omitempty"`
 	Oportunidades         *uint32  `json:"oportunidades,omitempty"`
-	Precio                *float64 `json:"precio,omitempty"`
+	Precio                *float32 `json:"precio,omitempty"`
 }
 
 func (product *Plan) TableName() string {
